Webscrape: Email Sender
This is a simple program I learned that uses BeautifulSoup to parse a website and locate the title and price of our desired consumer product. From there, we run a series of functions that check periodically if the price on the website is lower than the amount we want to pay. If the price lowers then the user will get an email notification letting them know "Price Lowered - Buy!".